// packages/vite/src/codegen.ts
export type Manifest = Record<
  string,
  { file: string; src?: string; isEntry?: boolean; imports?: string[]; css?: string[] }
>;